Add files via upload
the bitroar.unity is the scene that has our roary placed and in that scene there should be two parents - one named main camera and one named player. the main camera i haven't done anything with but with player that is roary and when you select him the movement.unitypackage project should pop up in a window somewhere within unity with an assets folder which contains the roary movement animations, movement script, etc. If for some reason roary doesn't move whenever you hit play, navigate to the animation controller and reset the values. I also added the roary movement script in here seperately just to be safe. It is labeled RoaryMoving.cs